# openhands/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHands Agent Server",
    "description": "An HTTP server to create and manage AI agent conversations using the OpenHands SDK.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/server/exec": {
      "post": {
        "tags": [
          "server"
        ],
        "summary": "Execute a bash command via subprocess.",
        "operationId": "server_exec_server_exec_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CmdExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CmdExecResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/server/fs/upload": {
      "post": {
        "tags": [
          "server"
        ],
        "summary": "Upload a file to the server filesystem.",
        "operationId": "fs_upload_server_fs_upload_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "dest_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Destination file path under FS_ROOT (default \"/\"). If exists, will overwrite. Intermediate directories will be created as needed.",
              "title": "Dest Path"
            },
            "description": "Destination file path under FS_ROOT (default \"/\"). If exists, will overwrite. Intermediate directories will be created as needed."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_fs_upload_server_fs_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/server/fs/download": {
      "get": {
        "tags": [
          "server"
        ],
        "summary": "Fs Download",
        "operationId": "fs_download_server_fs_download_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "File path under FS_ROOT.",
              "title": "Path"
            },
            "description": "File path under FS_ROOT."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_fs_upload_server_fs_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_fs_upload_server_fs_upload_post"
      },
...
      "CmdExecRequest": {
        "properties": {
          "command": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Command",
            "description": "Executable and args; no shell."
          },
          "timeout": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Timeout",
            "default": 30
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd",
            "description": "Working directory under FS_ROOT."
          },
          "env": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Env"
          },
          "max_output_bytes": {
            "type": "integer",
            "maximum": 10000000.0,
            "minimum": 1.0,
            "title": "Max Output Bytes",
            "default": 1000000
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "CmdExecRequest"
      },
      "CmdExecResponse": {
        "properties": {
          "stdout": {
            "type": "string",
            "title": "Stdout"
          },
          "stderr": {
            "type": "string",
            "title": "Stderr"
          },
          "return_code": {
            "type": "integer",
            "title": "Return Code"
          },
          "timed_out": {
            "type": "boolean",
            "title": "Timed Out"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "stdout",
          "stderr",
          "return_code",
          "timed_out",
          "duration_ms"
        ],
        "title": "CmdExecResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Master-Key"
      }
    }
  }
}